FIX BUILD: toybox: disable egrep, fgrep, grep and iotop.
OORDER MATTERS!

For reference see:
https://github.com/nvertigo/android_external_toybox/commit/cd40f1d10b9eb648aac44c83632e7f21e7430a3b

This breaks the rom by breaking toybox:

I two modules of the build process try to install /system/bin/XX, the
consequences are quite expectable: the LAST write wins. So far so good.

If the later of the two modules is not creating a regular file, but a
symblic link with "ln -sf" "unlink();" is called and the creation of the
symlink does not fail.

The other side around is expectable as well: if a module tries to
install the XX binary, when XX is already in place as symlink pointing
to toybox, "open();" is called. Depending on the way "install" is called
(or cp - for this matter), this just writing to /system/bin/XX.

Since XX is just a reference to toybox, toybox is overwritten.

For iotop (i.e.) this is currently happening:
1. The toolbox symlinks are setup pointing to toybox, which is not
     installed at this moment. iotop symlink pointing to toybox is in
     place.
2. toolbox binary is installed.
3. iotop (standalone command) is written to iotop symlink.
4. toybox was overwritten by iotop in 3.

Why is this happening?
If you've initializedna clean BUILD_TOP with only oneplus3t manifest,
there are two runs of toybox-instlist (host binary build in toybox,
sourcing generated/newtoys.h for generating symlink list). The second
one AFTER all standalone binaries were installed: so the standalone
binaries are replaced (due to "unlink();" prior to creating the
symlink).

So we don't have the intended standalone binaries, but this is hard to know,
because toybox is serving the requests for the command.

On different build environments this order has changed suddenly - so
toybox gets overwritten and having iotop or grep instead of toybox leads
up to all kinds of funny side effects. In deed, all shell scripts
relaying on shell comands served by toybox and every app doing "exec();"
on these commands, will crash as a matter of fact.

Symptons of this include, but don't limit to:
  * terminal gets started, but never delivers a shell prompt.
  * Substratum crashing on startup.
  * some subsystems are not initialized correctly (i.e. lte/ims).
  * ...

Why is the build order changing?
On the fresh and only for op3/T created build tree some more devices
were added. For my build environment I added bacon, gemini, marlin and
chagallwifi. After adding chagallwifi, the build order regarding toybox
has changed.

As explained en detail in the referenced commit message, there are some
sources to populate /system/bin. I decided to disable grep and iotop in
toybox, because I think, if a standalone binary's sourcecode has been
pulled to the build tree, it has been put in to be used... Maybe this is
a false assumption, but since no design- or configuration-concept is
visible for me, this is the best guess.

The other poss8ble fix would be deleting iotop's source dir, and
configure toolbox not to build ef-grep. But this would mean, that sorce
code that is not used, was added by intend to increase build time and
overwite it in the end, relaying on something fragile like the order of
building packages.